+++ rust/candid_parser/src/bindings/rust.rs
@@ -4,8 +4,10 @@ use crate::{
     syntax::{self, IDLActorType, IDLMergedProg, IDLType},
     Deserialize,
 };
-use candid::pretty::utils::*;
-use candid::types::{Field, Function, Label, SharedLabel, Type, TypeEnv, TypeInner};
+use candid::types::{
+    internal::TypeKey, Field, Function, Label, SharedLabel, Type, TypeEnv, TypeInner,
+};
+use candid::{pretty::utils::*, types::ArgType};
 use convert_case::{Case, Casing};
 use pretty::RcDoc;
 use serde::Serialize;
@@ -15,7 +17,7 @@ use std::collections::{BTreeMap, BTreeSet};
 const DOC_COMMENT_PREFIX: &str = "/// ";
 
 /// Maps the names generated during nominalization to the original syntactic types.
-type GeneratedTypes<'a> = BTreeMap<String, &'a IDLType>;
+type GeneratedTypes<'a> = BTreeMap<TypeKey, &'a IDLType>;
 
 #[derive(Default, Deserialize, Clone, Debug)]
 pub struct BindingConfig {
@@ -239,9 +241,15 @@ impl<'a> State<'a> {
                 .map(|(k, v)| (k.clone(), v.clone()))
                 .collect(),
         );
-        let src = candid::pretty::candid::pp_init_args(&env, &[src.clone()])
-            .pretty(80)
-            .to_string();
+        let src = candid::pretty::candid::pp_named_init_args(
+            &env,
+            &[ArgType {
+                name: None,
+                typ: src.clone(),
+            }],
+        )
+        .pretty(80)
+        .to_string();
         let match_path = self.state.config_source.get("use_type").unwrap().join(".");
         let test_name = use_type.replace(|c: char| !c.is_ascii_alphanumeric(), "_");
         let body = format!(
@@ -286,7 +294,7 @@ fn test_{test_name}() {{
                 Text => str("String"),
                 Reserved => str("candid::Reserved"),
                 Empty => str("candid::Empty"),
-                Var(ref id) => self.pp_var(id, is_ref),
+                Var(ref id) => self.pp_var(id.as_str(), is_ref),
                 Principal => str("Principal"),
                 Opt(ref t) => self.pp_opt(t, is_ref),
                 // It's a bit tricky to use `deserialize_with = "serde_bytes"`. It's not working for `type t = blob`
@@ -498,13 +506,14 @@ fn test_{test_name}() {{
 
     fn pp_defs(&mut self, def_list: &'a [&'a str]) -> RcDoc<'a> {
         let mut res = Vec::with_capacity(def_list.len());
-        for id in def_list {
+        for &id in def_list {
             let old = self.state.push_state(&StateElem::Label(id));
             if self.state.config.use_type.is_some() {
                 self.state.pop_state(old, StateElem::Label(id));
                 continue;
             }
-            let ty = self.state.env.find_type(id).unwrap();
+            let type_key: TypeKey = id.into();
+            let ty = self.state.env.find_type(&type_key).unwrap();
             let name = self
                 .state
                 .config
@@ -515,7 +524,7 @@ fn test_{test_name}() {{
             let syntax = self.prog.lookup(id);
             let syntax_ty = syntax
                 .map(|b| &b.typ)
-                .or_else(|| self.generated_types.get(*id).map(|t| &**t));
+                .or_else(|| self.generated_types.get(&type_key).map(|t| &**t));
             let docs = syntax
                 .map(|b| pp_docs(b.docs.as_ref()))
                 .unwrap_or(RcDoc::nil());
@@ -601,8 +610,18 @@ fn test_{test_name}() {{
         }
         lines(res.into_iter())
     }
-    fn pp_args<'b>(&mut self, args: &'b [Type], prefix: &'b str) -> RcDoc<'b> {
-        let tys = args.iter().enumerate().map(|(i, t)| {
+    fn pp_args<'b>(&mut self, args: &'b [ArgType], prefix: &'b str) -> RcDoc<'b> {
+        let doc = args.iter().enumerate().map(|(i, t)| {
+            let lab = t.name.clone().unwrap_or_else(|| format!("{prefix}{i}"));
+            let old = self.state.push_state(&StateElem::Label(&lab));
+            let res = self.pp_ty(&t.typ, true);
+            self.state.pop_state(old, StateElem::Label(&lab));
+            res
+        });
+        enclose("(", concat(doc, ","), ")")
+    }
+    fn pp_rets<'b>(&mut self, rets: &'b [Type], prefix: &'b str) -> RcDoc<'b> {
+        let tys = rets.iter().enumerate().map(|(i, t)| {
             let lab = format!("{prefix}{i}");
             let old = self.state.push_state(&StateElem::Label(&lab));
             let res = self.pp_ty(t, true);
@@ -611,14 +630,11 @@ fn test_{test_name}() {{
         });
         enclose("(", concat(tys.into_iter(), ","), ")")
     }
-    fn pp_rets<'b>(&mut self, rets: &'b [Type]) -> RcDoc<'b> {
-        self.pp_args(rets, "ret")
-    }
     fn pp_ty_func<'b>(&mut self, f: &'b Function) -> RcDoc<'b> {
         let lab = StateElem::TypeStr("func");
         let old = self.state.push_state(&lab);
         let args = self.pp_args(&f.args, "arg");
-        let rets = self.pp_rets(&f.rets);
+        let rets = self.pp_rets(&f.rets, "ret");
         let modes = candid::pretty::candid::pp_modes(&f.modes);
         let res = args
             .append(" ->")
@@ -668,7 +684,7 @@ fn test_{test_name}() {{
             .iter()
             .enumerate()
             .map(|(i, arg)| {
-                let lab = format!("arg{i}");
+                let lab = arg.name.clone().unwrap_or_else(|| format!("arg{i}"));
                 let old = self.state.push_state(&StateElem::Label(&lab));
                 let name = self
                     .state
@@ -677,7 +693,7 @@ fn test_{test_name}() {{
                     .clone()
                     .unwrap_or_else(|| lab.clone());
                 self.state.update_stats("name");
-                let res = self.pp_ty(arg, true);
+                let res = self.pp_ty(&arg.typ, true);
                 self.state.pop_state(old, StateElem::Label(&lab));
                 (name, res)
             })
@@ -727,7 +743,7 @@ fn test_{test_name}() {{
                 .iter()
                 .enumerate()
                 .map(|(i, arg)| {
-                    let lab = format!("arg{i}");
+                    let lab = arg.name.clone().unwrap_or_else(|| format!("arg{i}"));
                     let old = self.state.push_state(&StateElem::Label(&lab));
                     let name = self
                         .state
@@ -736,7 +752,7 @@ fn test_{test_name}() {{
                         .clone()
                         .unwrap_or_else(|| lab.clone());
                     self.state.update_stats("name");
-                    let res = self.pp_ty(arg, true);
+                    let res = self.pp_ty(&arg.typ, true);
                     self.state.pop_state(old, StateElem::Label(&lab));
                     (name, res.pretty(LINE_WIDTH).to_string())
                 })
@@ -794,7 +810,9 @@ pub fn emit_bindgen(
     let def_list = if let Some(actor) = &actor {
         chase_actor(&env, actor).unwrap()
     } else {
-        env.0.iter().map(|pair| pair.0.as_ref()).collect::<Vec<_>>()
+        env.to_sorted_iter()
+            .map(|pair| pair.0.as_str())
+            .collect::<Vec<_>>()
     };
     let recs = infer_rec(&env, &def_list).unwrap();
     let mut state = State {
@@ -1009,11 +1027,12 @@ impl<'b> NominalState<'_, 'b> {
                         &TypeInner::Record(fs.to_vec()).into(),
                         syntax,
                     );
+                    let new_var: TypeKey = new_var.into();
                     env.0.insert(new_var.clone(), ty);
                     if let Some(syntax) = syntax {
                         self.generated_types.insert(new_var.clone(), syntax);
                     }
-                    TypeInner::Var(new_var)
+                    TypeInner::Var(new_var.into())
                 }
             }
             TypeInner::Variant(fs) => {
@@ -1054,11 +1073,12 @@ impl<'b> NominalState<'_, 'b> {
                         &TypeInner::Variant(fs.to_vec()).into(),
                         syntax,
                     );
+                    let new_var: TypeKey = new_var.into();
                     env.0.insert(new_var.clone(), ty);
                     if let Some(syntax) = syntax {
                         self.generated_types.insert(new_var.clone(), syntax);
                     }
-                    TypeInner::Var(new_var)
+                    TypeInner::Var(new_var.into())
                 }
             }
             TypeInner::Func(func) => match path.last() {
@@ -1071,7 +1091,7 @@ impl<'b> NominalState<'_, 'b> {
                             .into_iter()
                             .enumerate()
                             .map(|(i, arg)| {
-                                let lab = format!("arg{i}");
+                                let lab = arg.name.clone().unwrap_or_else(|| format!("arg{i}"));
                                 let old = self.state.push_state(&StateElem::Label(&lab));
                                 let idx = if i == 0 {
                                     "".to_string()
@@ -1079,10 +1099,13 @@ impl<'b> NominalState<'_, 'b> {
                                     i.to_string()
                                 };
                                 path.push(TypePath::Func(format!("arg{idx}")));
-                                let ty = self.nominalize(env, path, &arg, None);
+                                let ty = self.nominalize(env, path, &arg.typ, None);
                                 path.pop();
                                 self.state.pop_state(old, StateElem::Label(&lab));
-                                ty
+                                ArgType {
+                                    name: arg.name.clone(),
+                                    typ: ty,
+                                }
                             })
                             .collect(),
                         rets: func
@@ -1120,8 +1143,8 @@ impl<'b> NominalState<'_, 'b> {
                         &TypeInner::Func(func.clone()).into(),
                         None,
                     );
-                    env.0.insert(new_var.clone(), ty);
-                    TypeInner::Var(new_var)
+                    env.0.insert(new_var.clone().into(), ty);
+                    TypeInner::Var(new_var.into())
                 }
             },
             TypeInner::Service(serv) => match path.last() {
@@ -1152,8 +1175,8 @@ impl<'b> NominalState<'_, 'b> {
                         &TypeInner::Service(serv.clone()).into(),
                         None,
                     );
-                    env.0.insert(new_var.clone(), ty);
-                    TypeInner::Var(new_var)
+                    env.0.insert(new_var.clone().into(), ty);
+                    TypeInner::Var(new_var.into())
                 }
             },
             TypeInner::Class(args, ty) => {
@@ -1168,10 +1191,13 @@ impl<'b> NominalState<'_, 'b> {
                             let elem = StateElem::Label("init");
                             let old = self.state.push_state(&elem);
                             path.push(TypePath::Init);
-                            let ty = self.nominalize(env, path, arg, None);
+                            let ty = self.nominalize(env, path, &arg.typ, None);
                             path.pop();
                             self.state.pop_state(old, elem);
-                            ty
+                            ArgType {
+                                name: arg.name.clone(),
+                                typ: ty,
+                            }
                         })
                         .collect(),
                     self.nominalize(env, path, ty, syntax_ty),
@@ -1192,12 +1218,17 @@ impl<'b> NominalState<'_, 'b> {
         prog: &'b IDLMergedProg,
     ) -> (TypeEnv, Option<Type>) {
         let mut res = TypeEnv(Default::default());
-        for (id, ty) in self.state.env.0.iter() {
-            let elem = StateElem::Label(id);
+        for (id, ty) in self.state.env.to_sorted_iter() {
+            let elem = StateElem::Label(id.as_str());
             let old = self.state.push_state(&elem);
-            let syntax = prog.lookup(id).map(|t| &t.typ);
-            let ty = self.nominalize(&mut res, &mut vec![TypePath::Id(id.clone())], ty, syntax);
-            res.0.insert(id.to_string(), ty);
+            let syntax = prog.lookup(id.as_str()).map(|t| &t.typ);
+            let ty = self.nominalize(
+                &mut res,
+                &mut vec![TypePath::Id(id.to_string())],
+                ty,
+                syntax,
+            );
+            res.0.insert(id.clone(), ty);
             self.state.pop_state(old, elem);
         }
         let actor = actor