+++ rust/candid_parser/src/syntax/pretty.rs
@@ -5,7 +5,9 @@ use crate::{
         candid::{pp_docs, pp_label_raw, pp_modes, pp_text},
         utils::{concat, enclose, enclose_space, ident, kwd, lines, str, INDENT_SPACE, LINE_WIDTH},
     },
-    syntax::{Binding, FuncType, IDLActorType, IDLMergedProg, IDLType, PrimType, TypeField},
+    syntax::{
+        Binding, FuncType, IDLActorType, IDLArgType, IDLMergedProg, IDLType, PrimType, TypeField,
+    },
 };
 
 fn pp_ty(ty: &IDLType) -> RcDoc {
@@ -94,13 +96,21 @@ fn pp_method(func: &FuncType) -> RcDoc {
         .nest(INDENT_SPACE)
 }
 
-fn pp_args(args: &[IDLType]) -> RcDoc {
-    let doc = concat(args.iter().map(pp_ty), ",");
+fn pp_args(args: &[IDLArgType]) -> RcDoc {
+    let args = args.iter().map(|arg| {
+        if let Some(name) = &arg.name {
+            pp_text(name).append(kwd(" :")).append(pp_ty(&arg.typ))
+        } else {
+            pp_ty(&arg.typ)
+        }
+    });
+    let doc = concat(args, ",");
     enclose("(", doc, ")")
 }
 
 fn pp_rets(rets: &[IDLType]) -> RcDoc {
-    pp_args(rets)
+    let doc = concat(rets.iter().map(pp_ty), ",");
+    enclose("(", doc, ")")
 }
 
 fn pp_service(methods: &[Binding]) -> RcDoc {
@@ -123,7 +133,7 @@ fn pp_service_methods(methods: &[Binding]) -> RcDoc {
     enclose_space("{", doc, "}")
 }
 
-fn pp_class<'a>(args: &'a [IDLType], t: &'a IDLType) -> RcDoc<'a> {
+fn pp_class<'a>(args: &'a [IDLArgType], t: &'a IDLType) -> RcDoc<'a> {
     let doc = pp_args(args).append(" ->").append(RcDoc::space());
     match t {
         IDLType::ServT(ref serv) => doc.append(pp_service_methods(serv)),