+++ rust/candid_parser/tests/parse_type.rs
@@ -5,6 +5,7 @@ use candid_parser::types::IDLProg;
 use candid_parser::typing::{check_file, check_prog};
 use goldenfile::Mint;
 use std::io::Write;
+use std::panic::AssertUnwindSafe;
 use std::path::Path;
 
 #[test]
@@ -132,7 +133,7 @@ fn compiler_test(resource: &str) {
             {
                 match filename.file_name().unwrap().to_str().unwrap() {
                     "unicode.did" | "escape.did" => check_error(
-                        || motoko::compile(&env, &actor, &prog),
+                        AssertUnwindSafe(|| motoko::compile(&env, &actor, &prog)),
                         "not a valid Motoko id",
                     ),
                     _ => {